+++ src/CL/Event.h
@@ -11,7 +11,7 @@ namespace CL
     {
         static const int MAX_ID_COUNT = 16;
         
-        long _ids[MAX_ID_COUNT];
+        int _ids[MAX_ID_COUNT];
         size_t _count;
 
         public:
@@ -23,7 +23,7 @@ namespace CL
         Event operator | (const Event& other) const;
         Event& operator = (const Event& other);
         const size_t get_id_count() const;
-        const long* get_ids() const;        
+        const int* get_ids() const;        
     };
 
 }
